backend/payments/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import PaymentTransactionViewSet, PayoutViewSet

router = DefaultRouter()
router.register(r"transactions", PaymentTransactionViewSet, basename="payment-transaction")
router.register(r"payouts", PayoutViewSet, basename="payout")

urlpatterns = [
    path("", include(router.urls)),
]

