backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-dotenv>=1.0
dj-database-url>=2.2
celery>=5.3
redis>=5.0
drf-spectacular>=0.27

